РБДиП/projectLaba3/refactored/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
python-dotenv
Pillow
Werkzeug